+++ CHANGELOG.md
@@ -9,9 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Forge
 
+#### Changed
+
+- Minimal supported `Scarb` version is now `2.10.0` (updated from `2.9.1`)
+
 #### Removed
 
 - Possibility to use `#[available_gas]` with unnamed argument. Use named arguments instead, e.g. `#[available_gas(l2_gas: 5)]`.
+- The deprecated command `snforge init`. Use `snforge new` to initialize new `Forge` projects
 
 ## [0.49.0] - 2025-09-03